less buggy: need to remove incorrect monitor guards
Some actions, which need not be guarded by the monitor, are being
guarded, we must get rid of them, otherwise the composition fails. When
those extra guards are removed, the composition should succeed (tried
this once by hand, but should be true in general).
Note, also fix the stopping conditions (we can also use them to get rid
of the duplication caused by randomly selecting states to be merged!)